requirements.txt
mcp[cli]>=0.1.0
anthropic>=0.40.0
google-api-python-client>=2.115.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pydantic>=2.5.3
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
rich>=13.7.0
tiktoken>=0.5.2
numpy>=1.26.3
scikit-learn>=1.4.0
